Plots a 3D representation of a cylinder bundle using plotly

Plots a 3D representation of a cylinder bundle from an object of class bundle as generated by simulate_bundle() using plotly.

Usage

plot3d(b, show_linear_mesh = FALSE)

Arguments

b

An object of class bundle as generated by simulate_bundle().

show_linear_mesh

A logical value indicating whether the linear mesh of each cylinder should be displayed. Defaults to FALSE for computational efficiency.

Value

An HTML widget of class plotly::plotly storing the 3D visualization of the cylinder bundle.

Examples

density <- 0.9
voxel_size <- 0.000005 # 5 micrometers
out <- simulate_bundle(density, voxel_size)
plot3d(out)