Plots a 3D representation of a cylinder bundle using rgl

Plots a 3D representation of a cylinder bundle from an object of class bundle as generated by simulate_bundle() using rgl.

Usage

# S3 method for bundle
plot3d(x, widget = FALSE, ...)

Arguments

x

An object of class bundle as generated by simulate_bundle().

widget

A logical value indicating whether to return an HTML widget for reporting. Defaults to FALSE.

...

Additional arguments to be passed to the rgl::plot3d(). Not used here.

Value

If widget is TRUE, an HTML widget is returned. Otherwise, nothing.

Examples

density <- 0.9
voxel_size <- 0.000010 # 10 micrometers
out <- simulate_bundle(density, voxel_size)
rgl::plot3d(out, widget = TRUE)

3D plot