Sphere compartment class

A class to model restricted diffusion in a sphere.

Methods

Public methods

SphereCompartment$new()

SphereCompartment$get_signal()

SphereCompartment$clone()

Method new()

Instantiates a new sphere compartment.

Usage

SphereCompartment$new(
  radius,
  diffusivity,
  model = c("soderman", "callaghan", "stanisz", "neuman", "vangelderen")
)

Arguments

radius

A numeric value specifying the radius of the sphere in meters.

diffusivity

A numeric value specifying the diffusivity within the sphere in m\(^2\).s\(^{-1}\).

model

A character string specifying the radial model to use. Choices are "soderman", "callaghan", "stanisz", "neuman", and "vangelderen". Defaults to "soderman".

Returns

An instance of the SphereCompartment class.

Method get_signal()

Computes the signal attenuation predicted by the model.

Usage

SphereCompartment$get_signal(
  small_delta,
  big_delta,
  G,
  echo_time = NULL,
  n_max = 20L,
  m_max = 50L
)

Arguments

small_delta

A numeric value specifying the duration of the gradient pulse in seconds.

big_delta

A numeric value specifying the duration between the gradient pulses in seconds.

G

A numeric value specifying the strength of the gradient in T.m\(^{-1}\).

echo_time

A numeric value specifying the echo time in seconds.

n_max

An integer value specifying the maximum order of the Bessel function. Defaults to 20L.

m_max

An integer value specifying the maximum number of extrema for the Bessel function. Defaults to 50L.

Returns

A numeric value storing the predicted signal attenuation.

Examples

sphereComp <- SphereCompartment$new(
  radius = 1e-6,
  diffusivity = 2.0e-9,
  model = "soderman"
)
sphereComp$get_signal(
  small_delta = 0.03,
  big_delta = 0.03,
  G = 0.040
)

Method clone()

The objects of this class are cloneable with this method.

Usage

SphereCompartment$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `SphereCompartment$get_signal`
## ------------------------------------------------

sphereComp <- SphereCompartment$new(
  radius = 1e-6,
  diffusivity = 2.0e-9,
  model = "soderman"
)
sphereComp$get_signal(
  small_delta = 0.03,
  big_delta = 0.03,
  G = 0.040
)
#> [1] 0.974496